// src/api/models/models-ShipModule/shipModuleSymbol.ts
/**
 * Generated by orval v8.0.0-rc.1 🍺
 * Do not edit manually.
 * SpaceTraders API
 * SpaceTraders is an open-universe game and learning platform that offers a set of HTTP endpoints to control a fleet of ships and explore a multiplayer universe.

The API is documented using [OpenAPI](https://github.com/SpaceTradersAPI/api-docs). You can send your first request right here in your browser to check the status of the game server.

```json http
{
  "method": "GET",
  "url": "https://api.spacetraders.io/v2",
}
```

Unlike a traditional game, SpaceTraders does not have a first-party client or app to play the game. Instead, you can use the API to build your own client, write a script to automate your ships, or try an app built by the community.

We have a [Discord channel](https://discord.com/invite/jh6zurdWk5) where you can share your projects, ask questions, and get help from other players.



 * OpenAPI spec version: 2.3.0
 */

/**
 * The symbol of the module.
 */
export type ShipModuleSymbol =
  (typeof ShipModuleSymbol)[keyof typeof ShipModuleSymbol];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ShipModuleSymbol = {
  MODULE_MINERAL_PROCESSOR_I: 'MODULE_MINERAL_PROCESSOR_I',
  MODULE_GAS_PROCESSOR_I: 'MODULE_GAS_PROCESSOR_I',
  MODULE_CARGO_HOLD_I: 'MODULE_CARGO_HOLD_I',
  MODULE_CARGO_HOLD_II: 'MODULE_CARGO_HOLD_II',
  MODULE_CARGO_HOLD_III: 'MODULE_CARGO_HOLD_III',
  MODULE_CREW_QUARTERS_I: 'MODULE_CREW_QUARTERS_I',
  MODULE_ENVOY_QUARTERS_I: 'MODULE_ENVOY_QUARTERS_I',
  MODULE_PASSENGER_CABIN_I: 'MODULE_PASSENGER_CABIN_I',
  MODULE_MICRO_REFINERY_I: 'MODULE_MICRO_REFINERY_I',
  MODULE_ORE_REFINERY_I: 'MODULE_ORE_REFINERY_I',
  MODULE_FUEL_REFINERY_I: 'MODULE_FUEL_REFINERY_I',
  MODULE_SCIENCE_LAB_I: 'MODULE_SCIENCE_LAB_I',
  MODULE_JUMP_DRIVE_I: 'MODULE_JUMP_DRIVE_I',
  MODULE_JUMP_DRIVE_II: 'MODULE_JUMP_DRIVE_II',
  MODULE_JUMP_DRIVE_III: 'MODULE_JUMP_DRIVE_III',
  MODULE_WARP_DRIVE_I: 'MODULE_WARP_DRIVE_I',
  MODULE_WARP_DRIVE_II: 'MODULE_WARP_DRIVE_II',
  MODULE_WARP_DRIVE_III: 'MODULE_WARP_DRIVE_III',
  MODULE_SHIELD_GENERATOR_I: 'MODULE_SHIELD_GENERATOR_I',
  MODULE_SHIELD_GENERATOR_II: 'MODULE_SHIELD_GENERATOR_II',
} as const;