// src/api/models/models-ShipMount/shipMountSymbol.ts
/**
 * Generated by orval v8.0.0-rc.0 🍺
 * Do not edit manually.
 * SpaceTraders API
 * SpaceTraders is an open-universe game and learning platform that offers a set of HTTP endpoints to control a fleet of ships and explore a multiplayer universe.

The API is documented using [OpenAPI](https://github.com/SpaceTradersAPI/api-docs). You can send your first request right here in your browser to check the status of the game server.

```json http
{
  "method": "GET",
  "url": "https://api.spacetraders.io/v2",
}
```

Unlike a traditional game, SpaceTraders does not have a first-party client or app to play the game. Instead, you can use the API to build your own client, write a script to automate your ships, or try an app built by the community.

We have a [Discord channel](https://discord.com/invite/jh6zurdWk5) where you can share your projects, ask questions, and get help from other players.



 * OpenAPI spec version: 2.3.0
 */

/**
 * Symbo of this mount.
 */
export type ShipMountSymbol =
  (typeof ShipMountSymbol)[keyof typeof ShipMountSymbol];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ShipMountSymbol = {
  MOUNT_GAS_SIPHON_I: "MOUNT_GAS_SIPHON_I",
  MOUNT_GAS_SIPHON_II: "MOUNT_GAS_SIPHON_II",
  MOUNT_GAS_SIPHON_III: "MOUNT_GAS_SIPHON_III",
  MOUNT_SURVEYOR_I: "MOUNT_SURVEYOR_I",
  MOUNT_SURVEYOR_II: "MOUNT_SURVEYOR_II",
  MOUNT_SURVEYOR_III: "MOUNT_SURVEYOR_III",
  MOUNT_SENSOR_ARRAY_I: "MOUNT_SENSOR_ARRAY_I",
  MOUNT_SENSOR_ARRAY_II: "MOUNT_SENSOR_ARRAY_II",
  MOUNT_SENSOR_ARRAY_III: "MOUNT_SENSOR_ARRAY_III",
  MOUNT_MINING_LASER_I: "MOUNT_MINING_LASER_I",
  MOUNT_MINING_LASER_II: "MOUNT_MINING_LASER_II",
  MOUNT_MINING_LASER_III: "MOUNT_MINING_LASER_III",
  MOUNT_LASER_CANNON_I: "MOUNT_LASER_CANNON_I",
  MOUNT_MISSILE_LAUNCHER_I: "MOUNT_MISSILE_LAUNCHER_I",
  MOUNT_TURRET_I: "MOUNT_TURRET_I",
} as const;